Database Modification
To support the "Create Listing" form I added the 'listings' table, I modified the 'book' table, and I modified the 'photo' table. There is no front and back photo field. Both the front of a book and the back of a book will be stored in the PHT_Image field. The ISBN, Edition and Author are now listed on the 'book' table. The 'listing' table has fields for type of sale and listing title.